Work on the rendering support.
Rendering the data of the report takes place in a class for that
purpose.  A rendering engine is supplied the report config and the
report data, which it uses to produce its output format.

Tabloid will ship a CSV, default HTML, JSON, and possibly PDF formatter.
New formatters can be easily added